Virtual Levels full implementation (#567)
* Renames sub zones into virtual levels, fixes river gen

* oops

* More sub zone to virtual level renamings

* Virtual Levels

* fixes

* Update shuttle.dm

* aaa

* Shuttles loaded into transit now properly work huzzah

* Update maps.dm

* proper side turf middle handling

* Update immovable_rod.dm

* allocation jump as argument, some definizations

* fxxxeadag

* Unit test zone

* a

* Removes Hilbert Hotel

* Update create_and_destroy.dm

* removes repair mapgen module

* Update mapping.dm

* fixes

* as anything

* JMP buttons in debug

* Virtual z improvements, multi-z improvements, code improvements

* Turns out you need virtual z as init argument because somethings may want it

* makes sure all turf types have proper arguments for init

* oops

* fixes overmap and transit related stuff

* Day night and weather properly cleans up on deleting a map zone

* Update mice_migration.dm

+++ code/datums/components/transit_handler.dm
@@ -1,21 +1,25 @@
 /datum/component/transit_handler
 	/// Our transit instance
 	var/datum/transit_instance/transit_instance
+	/// Time until we'll consider stranding a thing, they need to be in transit for some time until we delete them
+	var/time_until_strand = 0
 
 /datum/component/transit_handler/Initialize(datum/transit_instance/transit_instance_)
 	if(!ismovable(parent))
 		return COMPONENT_INCOMPATIBLE
 	transit_instance = transit_instance_
+	time_until_strand = world.time + 1 SECONDS
 	transit_instance.affected_movables[parent] = TRUE
 	RegisterSignal(parent, COMSIG_MOVABLE_MOVED, .proc/on_parent_moved)
 
 /datum/component/transit_handler/proc/on_parent_moved(atom/movable/source, atom/old_loc, Dir, Forced)
-	var/turf/new_location = get_turf(parent)
+	if(QDELETED(source))
+		return
+	var/turf/new_location = get_turf(source)
 	if(!istype(new_location, /turf/open/space/transit))
 		qdel(src)
-	transit_instance.MovableMoved(parent)
+	transit_instance.movable_moved(source, time_until_strand)
 
-/datum/component/transit_handler/Destroy()
+/datum/component/transit_handler/UnregisterFromParent()
 	UnregisterSignal(parent, COMSIG_MOVABLE_MOVED)
 	transit_instance.affected_movables -= parent
-	return ..()
